+++ src/contentTypeConfig.js
@@ -66,6 +66,8 @@ export const DEFAULT_FIELDS = [
   'child_count',
   '"parent_id": parent_content_reference[0]->railcontent_id',
   '"grandparent_id": parent_content_reference[1]->railcontent_id',
+  'live_event_start_time',
+  'live_event_end_time',
 ]
 
 // these are identical... why
@@ -349,7 +351,7 @@ const lessonRecentTypes = [
   'routine',
 ]
 
-const parentRecentTypes = [
+export const parentRecentTypes = [
   ...individualLessonsTypes,
   'skill-pack',
   ...entertainmentLessonTypes,
@@ -403,6 +405,18 @@ export let contentTypeConfig = {
       `"badge_logo" : ${contentAwardField}.logo.asset->url`,
     ],
     includeChildFields: true,
+  },
+  'hierarchy-data': {
+    fields: [
+      'railcontent_id',
+      '"assignments": assignment[]{railcontent_id}',
+      '"metadata": { brand, "type": _type, "parent_id":  coalesce(parent_content_data[0].id, 0) }',
+      ],
+    childFields: [
+      'railcontent_id',
+      '"assignments": assignment[]{railcontent_id}',
+      '"metadata": { brand, "type": _type, "parent_id":  coalesce(parent_content_data[0].id, 0) }',
+    ],
   },
   song: {
     fields: ['album', 'soundslice', 'instrumentless', `"resources": ${resourcesField}`],
@@ -653,7 +667,10 @@ export let contentTypeConfig = {
     }`,
   ],
   'new-and-scheduled': {
-    fields: ['show_in_new_feed', isLiveField()],
+    fields: [
+      'show_in_new_feed',
+      isLiveField()
+    ],
   },
 }